requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.2.0
langchainhub
openai
chromadb
bs4
tiktoken
unstructured
python-docx
pypdf
pandas
python-dotenv